requirements.txt
streamlit
pandas
plotly
biopython
py3Dmol
stpyvmol
